+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.7.0 / 2021-03-07
+
+- [ENHANCEMENT] Add serverless deployment
+- [ENHANCEMENT] Add default config for custom deployments
+
 ## v1.6.0 / 2021-03-05
 
 - [FEATURE] Add cluster alarms